offer selective override
